Fix subject update error if assignment exists
Previously the subject update would replace the original entry which
would break the assignement-subject relation. That was prevented by ON
DELETE RESTRICT assignment FK constraint, however it would abort the
update all together and the new subject data would be rejected.

Switching to INSERT ... ON CONFLICT ... structure, updates the data,
while perserving the subject-assignment relation and without triggering the
ON DELETE RESTRICT constraint.